Removing `true_objective_metric_name` from early stopping strategies (#2214)
Summary:

This commit removes `true_objective_metric_name` from early stopping strategies. 

`true_objective_metric_name` is used when the early stopping strategy works on a different metric than any of the objectives of the experiment. In this case however, I’d expect the “true objective” to show up in the experiments optimization config, so we wouldn’t need to keep track on it on the ESS. 

The only case I can think of where `true_objective_metric_name` would be required is when the “true objective” is not an actual objective on the experiment. Notably, early stopping strategies were historically registered as objectives of an experiment, and the "actual" optimization objectives were registered as tracking metrics. However, this setup was changed on the IGML side with D46800507 following the discussion in [this post](https://fb.workplace.com/groups/822676228356687/permalink/1252543232036649/). 

Therefore, it seems we can remove `true_objective_metric_name`.

Differential Revision: D54019501